feat: naive Jupyter provider
For now, the provider naively provides and resolves a static list of
Colab servers. Additionally, the RP connection information is provided
as a hardcoded configuration until we have assignment logic implemented.
Both of these items are flagged with TODOs.

Other minor changes:

- Deleted the temporary `tmp.unit.test.ts` since we now have a _real_
  unit test. As per https://github.com/googlecolab/colab-vscode/pull/4.
- Updated the `test:unittests` script to quote the glob pattern that's
  used. This was a bit of a fun one to figure out. In short, when left
as a naked glob, the shell will expand it. on Linux `sh` will not enable
globstar by default. This was causing the test files to not both be
resolved (prior to deleting the `tmp` one). This is most effectively
solved by quoting the glob which in turn relies on the glob tool used by
the tool (`mocha`) to consistently expand the star. [This blog
post](https://medium.com/@jakubsynowiec/you-should-always-quote-your-globs-in-npm-scripts-621887a2a784)
covers this in detail.
- Bumped the VS Code version.
- Added a dependency on `chai` for assertions and `sinon` for spies,
  stubs and mocks. Across the extension ecosystem and TS/JS in general,
these are the (current) most popular choices.