ci: configure unit test automation
For now, a temporary unit test file has been created to bootstrap the
workflow. It will be deleted when the first real unit test(s) are added.

Alongside this change, snapping to a single node version for testing as
opposed to the matrix strategy. Until we have a good reason, no sense
testing against other versions.

Since we're adding a `mocha` script, updated the relevant package(s).