chore: Correct config settings Resources misspelling
Corrects misspelling of "Resoruces" to "Resources" in the settings config.
This misspelling is distracting and is relied on when grabbing the
config properties in the extension code, we can prevent strange
reliance on misspelled words by fixing this early.

Additionally moved config related strings to an enum instead of hard
coding them in place.