test: establish unit test pattern with VS Code.
Unit testing a VS Code extension is super tricky! I spent a bit longer on this than I care to admit, but I think I've found a pattern which makes sense. I'll do my best to explain that here and link out to the resources which helped influence the approach.

The crux of the problem is `vscode` is not a normal node dependency. Rather, it's the _engine_ that runs the application. This means that we can't just import `vscode` into our tests and expect it to work. It's only available when VS Code is running, or the integration test runner is driving the tests. I like unit tests, a lot, so I wanted to find a way to test the extension without having to run the entire application.

With a few sub-option branches, there are basically two options for achieving this. You can either (1) wrap the `vscode` API or (2) mock it. If you're really curious what these (and the various sub-options) look like, you can checkout the resources I found and used to shape my approach. I tried all of them (including swapping to Jest) 😫.

- https://gitlab.com/gitlab-org/gitlab-vscode-extension/-/snippets/3710659
- https://devblogs.microsoft.com/ise/testing-vscode-extensions-with-typescript/#option-2:-use-jest-manual-mocks
- https://app.studyraid.com/en/read/8400/231878/mocking-vs-code-api
- https://www.richardkotze.com/coding/unit-test-mock-vs-code-extension-api-jest

By injecting the entire `vscode` module, it not only makes test setup more clear and consistent, but also encourages a single point of injection for the `vscode` environment (engine). While this PR only adds the stubs needed to swap the Jupyter extension integration test to a unit test, subsequent PRs add more tests and further build out the `vscodeStub`.