feat: Adds custom fetch logic
Adds custom fetch logic to add proxy tokens to headers of all fetchs.
Because the original implementation in the Jupyter extension uses the
the 'node-fetch', I have opted to do the same as well since that will
likely have the least amount of edge cases in the future.

Because this requires us to rely on the jupyter extensions proposed API,
the package version has been downgraded to the latest version with the
"proposed" tag.