feat: Add a ccu info class that maintains ccu info
Adds a CCUInformation class that maintains the the ccu info. This is
The class maintains a set interval of 5 minutes that when information
changes updates the state in the class and emits an event to the
provider to notify it to rerun the jupyter server provider methods.

Several nuances with the CCUInformation class:
- it has a static initialize method since it relies on the client
  `ccuInfo` promise to be resolved, and we don't want to wait an unknown
  amount of time in the constructor.
- maintains the interval timer as it's own property to ensure that we
  can dispose of it properly.
- updates it's own `ccuInfo` property and emits a void event instead of
  emitting the `ccuInfo` value. This is so that there is a single
  central place for this, instead of managing `ccuInfo` in multiple
  places.