fix: pin node-fetch to specific version
This looks strange to me, but when running node-fetch in version 2.6.1
and 2.7.0 an error that absolute path is required. However, when testing
against the base package that the jupyter extension uses 2.6.7 that is
not the case. pinning the node-fetch version and extra code to create a
request with an absolute path.