Fix: add logic to get around node-fetch symbol bug
Node fetch has an interesting bug where symbols do not always match
requests. This seems to be a known issue but not prevalant enough to
warrant an issue, and there is a known workaround in [their
discussions](https://github.com/node-fetch/node-fetch/discussions/1598).

The Jupyter extension seems to get around throwing this typeError by
strategically type asserting to `any` and using their own `RequestInfo`
type.

Using a similar type strategy will likely make it more difficult for us
to debug issues in the future, since we will need to follow the type
pipeline from jupyter extension to the Colab extension back to the
jupyter extension and finally into node-fetch....

On the other hand using the workaround is ugly, but any issues that
occur with the fetch function we can still relatively easily debug
within our own extension.

I don't really know why sometimes running a fresh npm install pinned to
certain versions worked especially since if I left it for a day or so it
would no longer work, this lead to a lot of confusion/frustration on my
end.

Needless to say this fix will allow the fetch function to work properly
on all the `node-fetch` package versions that the Jupyter extension has
marked as a dependency.