feat: add support for storing the Colab Jupyter servers
On extension re-activation, VS Code reload, etc., we need to be able to populate the list of previously connected servers. While we could query for assignments every time, it's unecessary and slow. This storage class abstracts the complexity which will be leveraged by the re-worked Jupyter server provider in subsequent PRs.