fix: zod-ify auth storage
Bye-bye gnarly type-guards!

Currently only supports a single session, since we only ever need the one scope. Despite this, the implementation is designed to be extensible to multiple sessions in the future (stores an array of sessions). We are likely to do this if and when we support Drive-specific functionality.