chore: Add eslint rule for tsdoc
add tsdoc eslint rule and apply fix for tsdoc linter warnings

All of these fixes were applied manually.
There may be a way to at least apply some of these changes automatically
with something like Prettier. Will look into that in, and apply a
different patch if that is possible.