feat: abstract task running logic
The same robust synchronous task running logic is needed in multiple places
(keep-alive and CCU info). This centralizes it into a standalone class and
refactors the existing classes to leverage it.

I also took the chance to fix an issue with the `CcuInformation` class (renamed
as `CcuInformationManager`) where we were exposing `onDidChangeCcuInfo:
EventEmitter<void>;` instead of simply the `vscode.Event`. Additionally took a
pass at improving test clarity + behavioural coverage.