feat: auto-increment the server placeholder
Follows standard file explorer semantics for duplicating files.

- If no server exists, placeholder is `Colab CPU`.
- If `Colab CPU` exists, placeholder is `Colab CPU (1)`.
- If `Colab CPU` and `Colab CPU (1)` exist, placeholder is `Colab CPU (2)`.
- If `Colab CPU` and `Colab CPU (2)` exist (because `1` was removed),
  placeholder is `Colab CPU (1)`.

The above should be easily understood looking at the tests.

Determining the number to use requires the `AssignmentManager`, so the functions
were moved to members of the `ServerPicker` class.

I also re-organized the file a bit so exported members are at the top.