feat: add a user-info accessor to the Colab Google API
As with the existing API surface... the Colab backend APIs are a little funky!
The team's in the process of standardizing, normalizing and cleaning up the API
surface. A big part of this is moving our APIs to One Platform (Google APIs, or
GAPIs for short). In the extension, we need to know if the user is a pro/pro+
subscriber or not. Our existing frontend gets this through some funky
side-loaded bits of information we don't have access to in the extension.
Therefore, we need to integrate with the new GAPI for retrieving this user info.

I've done my best to compartmentalize all the funky API intricacies into
`api.ts`, `transform`-ing and _normalizing_ across the two backends so the rest
of the client code can be blissfully unaware of the differences.

Currently, this API requires a Drive-scoped OAuth token. This is due to some
internal details that don't really matter, but it is a temporary requirement and
we'll remove it before launch (hence the TODO).