feat: change development target to sandbox
We're in the process of reducing the scopes required to hit the Colab
backend(s). Currently, they require `drive`. This change temporarily adds that
scope and configures the extension to hit the sandbox environment. A follow-up
change will extract all of this environment-based config to `.env` files with
the supporting `launch.json` entries.