fix: convert remainingTokens to a number
The Proto-defined field is an `Int64`. As the
[docs](https://protobuf.dev/programming-guides/json) outline, `Int64`
fields are JSON serialized as strings since they can eclipse the max
number size in JavaScript (`2^53 - 1`).

In practice, this value should never be more than 6,000 (the current max
allowed tokens). We can safely convert, though I did add some safety to
the refinement/validation largely for clarity.