refactor: centralize the definition of commands
An additional minor change in here is removing the steps for the _remove server_
command as there's only the one step.

In a following change, the commands need to be referenced in multiple places.
This is effectively a prefactor for that. But even if that change wasn't going
to happen, this is a good cleanup.