fix: await setting the assignment context key
As is, setting the context key as a fire and forget promise introduces potential
race conditions. I came across that when adding the `hasAssignedServer` method
which I need in a follow-up change for the notebook toolbar button. This change
is a fix to the async code, a prefactor for the afformentioned change and an
opportunity to cleanup the tests:

- All the stubbing on the storage was hard to read and simple to fake.
  Introduced a thin in-memory implementation to make the tests more readable and
  easier to work with.
- Minor naming improvements.
- Removed unecessary implementation-detail-specific stub assertions in favour of
  interacting with the public API.
- Restructured `hasAssignedServer` test(s) to instead validate the context key
  is set correctly everywhere it's modified (reconcile, assign, refresh, unassign).