feat: implement generic loopback server
This class with be leveraged in a downstream change for a local server OAuth2
flow.

The unit tests leverage a fake server to remain targetted and quick to run. We
should consider adding an integration test that spins up a real server. This
will be inherently slower, but will ensure that the server works as expected
in a real environment.