feat: warn of server removals
Included in this change is also a step to reconcile the servers before listing them. This ensures that the server list is always up to date when the Jupyter extension needs it.

Now when a removal is detected which was not initiated by the user, a warning is displayed to inform them of the change. The more likely case of automatic removal is when a user stops using VS Code, then comes back after some period of time (e.g. the next day) and Colab has pruned the server. In those cases, no notification is shown since we reconcile on activation before registering the `ColabJupyterServerProvider` which binds this event / dispatches the notification.