feat: only prompt relevant auth flows
On activation, we determine where we're running (web or desktop). The flows are obtained once and provied statically to the login flow.

In working on this change, interestingly enough I found that with port forwarding enabled, you can actually spin up a local server in GitHub Codespaces. Despite that, it makes more sense to only use the proxied approach in Web environments. It avoids the need to accept port forwarding, aligns with web standards and will produce a more secure auth road-bump in the proxied redirect.