fix: correct prod Colab API domain
While the webapp will redirect fine from colab.google.com, we're
returning a 200 with some JS to do that 🫨. To avoid the redirect we
can't easily follow (as if it were a 300), this change configures the
URL to the redirected endpoint.

Additionally, this change updates the GAPI endpoint to the new domain.