fix: GPU and TPU assignment parameter bug
`api.ts` really is starting to show some inconsitencies with the Colab API! It's a bit funky to say the least. This change uses the appropriate enum type (num vs. string) over the wire and documents why and where the conversion is necessary.

Tested manually and can assign across all the machine types.

Took the opportunity to rework the `client.unit.test.ts` matcher.

Also added the response body of failed requestse to the error message.