fix: issue Jupyter API calls to the server directly
These initial calls went in when we were developing against local instances of Colab. Now that we've got auth sorted and are using _real_ Colab, it's clear we need to issue requests directly to the server, not the Colab tunnel.

I've manually tested both unassigning a server and verifying that listing kernels for the keep-alive calls works as expected (though keep-alive is still not working since we needed to allow-list that endpoint, which is done but not rolled out).