chore: omit unecessary files from packaging
```sh
 INFO  Files included in the VSIX:
colab-0.0.1.vsix
├─ [Content_Types].xml
├─ extension.vsixmanifest
└─ extension/
   ├─ LICENSE.txt [11.09 KB]
   ├─ SECURITY.md [0.36 KB]
   ├─ package.json [4.48 KB]
   ├─ readme.md
   ├─ out/
   │  ├─ extension.js [575.98 KB]
   │  └─ auth/
   │     └─ media/
   │        └─ favicon.ico [71.36 KB]
   └─ resources/
      ├─ colab_grayscale_dark.svg [2.1 KB]
      └─ colab_grayscale_light.svg [2.1 KB]
```

For now, I'm fine with all builds going to `out/`. I'm considering a separate `dist/` directory in the future to help ensure only _production_ assets are being included in the extension. This is a reasonable first step towards that goal.

This change also makes one small minor cleanup to the esbuild.