refactor: toggle background tasks
There's a handful of coupled changes here, sorry there wasn't an easier way to do this. I was finding the lifecycle management of background tasks a bit murky. This was exacerbated when I tried to implement a more granular control over their execution depending on whether the user was signed in to use Colab or not. This prefactor introduces a new `Toggleable` class to help manage the state of these background tasks more easily, instead of starting them on construction.