feat: bind a language client for the latest server
There are two inputs that determine how a language client is bound/registered:

1. Does the user have a valid auth session?
2. What's the latest server?

This controller listens to those two events and manages the lifecycle of the corresponding server. Here we use `LatestCancelable` to do the `async` _spin up_ in response to the aforementioned sync events. To facilitate testing and keep the logging in a centralized spot, when constructed we listen to the `ServerChangeEvent` this class emits.