refactor: lift "has servers" context to provider
The `AssignmentManager` deals strictly with APIs surrounding assignments, and doesn't concern itself with auth. It takes in a `ColabClient` and uses it. If the user isn't auth'd, they're prompted to sign-in when that client goes to get a token. This background is important when you consider what it means to claim there are assigned servers. Imagine the user has a server, but signs out. Do they have a sever? What about in 5 minutes? 60? We can't know unless we're authorized to check (query the API).

The `ColabJupyterServerProvider` is the natural place for making this determination, since it provides whatever servers it can, given the current authorization. When it's unauth'd it provides `[]`. This makes it the authoritative voice of whether or not there are assigned servers, so setting the context makes the most sense there.

This change actually solves the issue that caused me to move the functionality. Before, if you had a server and re-loaded the extension, the `Colab: Remove Server` wouldn't be available, since the context wasn't re-evaluated. Sure we could have done some kind of _initialization_ but I think this lift makes more sense. It has the benefit of doing without the unecessary toggling indirection, something I make look to do where it makes sense for the other toggles. Though for those that are pure "are you auth'd" toggles, it's less of a smell.