refactor: simplify async toggling
I wasn't happy with `AsyncToggleable`. There were a few things I didn't like, but the fact that it is generic highlights the biggest (`<T extends Disposable>`). It conflated resource lifecycle with execution lifecycle. by making it singly-responsible for execution lifecycle, the code reads more more easily and it gives the flexibility to callers to control when and how their resources are managed. Gone is the imiplicit indirection.

This is achieved by adding `LatestCancelable` which simply looks after running one async task at a time by cancelling the in-flight one. I like this pattern quite a bit since it pushes callsites to look like `this.runner.run("my", "params")`. Even more-so, if we run it from a sync context as is often the case from sync event listeners, the `void` makes this _runner_ more obvious. We're fire-and-forgetting this async runs.