feat: implement the `ProxiedJupyterClient`
This client is a thin wrapper of the APIs the generated `JupyterClient`
exposes. It lazily constructs the underlying API objects and offers two
factory functions (`static`):

- `withStaticConnection` - creates a client with the server's current
  connection information.
- `withRefreshingConnection` - creates a client that uses the server's
  updated connection information (new token as it changes).

The former is useful for one-off queries while the later is ideal for
persistent / long-lived clients.

A downstream change will leverage this client extensively as we use the
`ContentsApi` to enable the file system provider.