fix: register the auth provider on initialize
I discovered this hidden bug, which doesn't have a noticeable user impact, while working on another feature. I saw that during extension activation the `Must call initialize() first.` error was being thrown. This was happening since VS Code will start trying to use the auth provider once it's registered, and we were registering it on construction instead of when `initialize` was called (when it becomes actually ready).