chore: migrate Jupyter definition to Open API 3
Jupyter servers expose 2, which has one notable limitation that's needed for downstream PRs which add support for mounting the Jupyter content directory as a file system in VS Code.

Swagger 2 does not support `oneOf`, yet the Jupyter API returns "one of" a few types for `contents`. This is made clear in its comment:

> The content, if requested (otherwise null). Will be an array if type is 'directory'.

As is, our only option is to manually modify the generated code, or write complex post-generation swaps which are brittle or break the type system.

I started with breaking the type system and hated it. So I auto-migrated the API definition to 3 using https://converter.swagger.io/ (official supported tool) and re-generated. Working E2E as expected.