feat: enable mount commands in Colab extension
Configuring this as an _experimental_ config. This helps communicate to users that this feature is new, may have rough edges and encourages them to provide feedback.

Initially launching with it off by default. I'll be posting across the various GitHub issues and discussions to let users know about this feature and gather feedback.

The two entrypoints with this changes are the command palette and notebook toolbar.